Lots of changes. Added new scene "FollowThatPig" where the camera follows the pig around, with a parallax background.
Big trick with the CPDim--needs the scene and view to have explicit and different values for this to work correctly. Needs some cleanup.

Also starting to further organize--though this needs a lot of work. Breaking out "images" and "animations", for instance.

Leaving a bit of mess everywhere, but it's compiling and running.